+++ ClassLibraries/Macross.Json.Extensions/CHANGELOG.md
@@ -20,7 +20,7 @@
   decorate an enum type with the options to use when serializing/deserializing
   its values.
 
-* Add `JsonTypeConverterAdapterFactory` for using `TypeConverter`s with
+* Added `JsonTypeConverterAdapter` for using `TypeConverter`s with
   System.Text.Json. ([#19](https://github.com/Macross-Software/core/pull/19)) 
 
 ## 2.0.0